r591
Новая система локализации шаблонов боковых панелей, пользовательских соглашений и т.д. Добавлены локализации оных на английский и украинский языки. Исправления ошибок и прочие изменения: - вторая боковая панель включена по-умолчанию; - включена панель отладки в администраторской панели; - исправление ошибки с сохранением кеша для разных языков; - исправление ошибки с выбором стандартного языка для форума; - исправление ошибки с генерацией списка текущих дней рождения; - исправление ошибки с иногда некорректной генерацией списка файлов в раздаче; - исправление ошибки с возможностью парсинга базы данных пользовательских email-адресов; - исправление ошибки с редиректом страниц на некоторых конфигурациях серверов (IIS); - исправление ошибки с выводом информации о текущем выбранном шаблоне/языке, если на форуме он всего один; - исправление кодировки кастомного поиска для браузера; - множественные правки в css, js, языковых переменных. Также были переименованы все файлы в папке /includes/ucp, для предотвращения дублирования текста в названии оных. Изменения в сами файлы не вносились. git-svn-id: https://torrentpier2.googlecode.com/svn/trunk@591 a8ac35ab-4ca4-ca47-4c2d-a49a94f06293

+++ upload/ajax/view_torrent.php
@@ -104,7 +104,7 @@ class torrent
 				}
 				array_deep($f['path'], 'clean_tor_dirname');
 
-				$length = isset($f['length']) ? (int) $f['length'] : 0;
+				$length = isset($f['length']) ? (float) $f['length'] : 0;
 				$subdir_count = count($f['path']) - 1;
 
 				if ($subdir_count > 0)
